+++ src/main/java/com/nisovin/shopkeepers/commands/arguments/PlayerShopkeeperFilter.java
@@ -20,8 +20,8 @@ public class PlayerShopkeeperFilter implements ArgumentFilter<Shopkeeper> {
 	public String getInvalidArgumentErrorMsg(CommandArgument<Shopkeeper> argument, String input, Shopkeeper value) {
 		if (input == null) input = "";
 		return TextUtils.replaceArgs(Settings.msgCommandShopkeeperArgumentNoPlayerShop,
-				"{argumentName}", argument.getName(),
-				"{argumentFormat}", argument.getFormat(),
+				"{argumentName}", argument.getRootArgument().getName(),
+				"{argumentFormat}", argument.getRootArgument().getFormat(),
 				"{argument}", input);
 	}
 }