+++ src/main/java/com/nisovin/shopkeepers/commands/lib/arguments/FirstOfArgument.java
@@ -42,14 +42,15 @@ public class FirstOfArgument extends CommandArgument<Pair<CommandArgument<?>, Ob
 		super(name);
 
 		// arguments:
-		Validate.notNull(arguments);
+		Validate.notNull(arguments, "Arguments is null!");
 		List<CommandArgument<?>> argumentsList = new ArrayList<>(arguments.size());
 		this.arguments = Collections.unmodifiableList(argumentsList);
 		for (CommandArgument<?> argument : arguments) {
-			Validate.notNull(argument);
+			Validate.notNull(argument, "Contained argument is null!");
+			argument.setParent(this);
 			argumentsList.add(argument);
 		}
-		Validate.isTrue(this.arguments.size() != 0, "No (valid) arguments given!");
+		Validate.isTrue(this.arguments.size() != 0, "No arguments given!");
 
 		// format:
 		if (joinFormats) {